WIP: File analyzer. Perhaps do metaprogrammatically to take advantage of Ruby's method understanding?
What are base objects? Base objects are:
- Object
- Methods on object
- Messages sent by that method, to other objects.

So we have 4 primitives: Object, Methods (of Object), Messages, and Recipients (other Objects). Should be simple enough to extract these. Extract Objects, search their text for Methods, search the text of those for Messages, for each message figure out who the Recipient is, for each Object add it to a global graph of Object (Method) -> Message -> Object -> .... Probably want objects for each one.

Probably want objects for relations as well... Or those can be inferred from the belongs_to relationships there? perhaps. The only ones that need to be drawn as arrows are Object(within a Method) -> Messages -> Objects.

Whoo